requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-docx==0.8.11
PyPDF2==3.0.1
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2